tests/wpt/tests/shadow-dom/reference-target/tentative/interestfor.tentative.html

<!DOCTYPE HTML>
<html>

<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/resources/testdriver-vendor.js"></script>
  <script src="/resources/testdriver-actions.js"></script>
  <script src="/wai-aria/scripts/aria-utils.js"></script>
  <script src="/html/semantics/interestfor/resources/invoker-utils.js"></script>
</head>

<body>
  <style>
      [interestfor] {
          interest-delay: 0s;
      }
  </style>
  <button id="toggle-button-1" interestfor="x-popover-1">Toggle the popover</button>
  <x-popover-1 id="x-popover-1" popover>
    <template shadowrootmode="open" shadowrootreferencetarget="popover">
      <div id="popover" popover>Popover content inside shadow root</div>
    </template>
  </x-popover-1>

  <button id="toggle-button-2" interestfor="x-popover-2">Toggle the popover</button>
  <x-popover-2 id="x-popover-2" popover></x-popover-2>
  <script>
    const customPopover2 = document.querySelector('x-popover-2');
    customPopover2.attachShadow({ mode: 'open', referenceTarget: 'popover' });
    customPopover2.shadowRoot.innerHTML = '<div id="popover" popover>Popover content inside shadow root</div>';
  </script>

  <button id="toggle-button-3">Toggle the popover</button>
  <x-popover-3 id="x-popover-3" popover>
    <template shadowrootmode="open" shadowrootreferencetarget="popover">
      <div id="popover" popover>Popover content inside shadow root</div>
    </template>
  </x-popover-3>

  <button id="#other-hover-target">Other hover target</button>

  <script>
    function testInterestFor(popoverId, buttonId, name) {
        promise_test(async function () {
        const xPopover = document.getElementById(popoverId);
        const popover = xPopover.shadowRoot.getElementById("popover");
        const otherHoverTarget = document.getElementById('#other-hover-target');

        let xPopoverShowCount = 0;
        xPopover.addEventListener('beforetoggle', (e) => {
          if (e.newState === "open")
            ++xPopoverShowCount;
        });
        let popoverShowCount = 0;
        popover.addEventListener('beforetoggle', (e) => {
          if (e.newState === "open")
            ++popoverShowCount;
        });

        const toggleButton = document.getElementById(buttonId);
        await hoverOver(toggleButton);

        assert_equals(xPopoverShowCount, 0, "host should not be toggled if valid reference target is set");
        assert_equals(popoverShowCount, 1, "inner element should be target of toggle if reference target is set");
        popover.hidePopover();

        // Setting referenceTarget to null means the host element is the target of the idref.
        xPopover.shadowRoot.referenceTarget = null;
        await hoverOver(otherHoverTarget); // Move mouse away first
        await hoverOver(toggleButton);
        assert_equals(xPopoverShowCount, 1, "host should be toggled if reference target is null");
        assert_equals(popoverShowCount, 1, "inner element should not be toggled if reference target is null");
        popover.hidePopover();

        // Setting referenceTarget to empty string means the idref is treated as invalid.
        xPopover.shadowRoot.referenceTarget = "";
        await hoverOver(otherHoverTarget); // Move mouse away first
        await hoverOver(toggleButton);
        assert_equals(xPopoverShowCount, 1, "no additional toggle on host if reference target is empty string");
        assert_equals(popoverShowCount, 1, "no additional toggle on inner element if reference target is empty string");
      }, name);
    }
    testInterestFor('x-popover-1', 'toggle-button-1', "Shadow root reference target works with interestfor attribute.");
    testInterestFor('x-popover-2', 'toggle-button-2', "Shadow root reference target works with interestfor attribute via options.");
    document.getElementById('toggle-button-3').interestForElement = document.getElementById('x-popover-3');
    testInterestFor('x-popover-3', 'toggle-button-3', "Shadow root reference target works with .interestForElement property.");
  </script>
</body>

</html>
